+++ src/onegov.gazette/HISTORY.rst
@@ -1,6 +1,10 @@
 Changelog
 ---------
-1.13.15 (2018-04-26)
+
+- Adds the number of rejections by user to the statistics.
+  [msom]
+
+1.12.15 (2018-04-26)
 ~~~~~~~~~~~~~~~~~~~~
 
 - Updates translation.